+++ src/TestCase/WPTestCase.php
@@ -80,6 +80,9 @@ class WPTestCase extends Unit
 {
     use WPTestCasePHPUnitMethodsTrait;
 
+    public static bool $beStrictAboutWpdbConnectionId = true;
+    private static ?string $wpdbConnectionId = null;
+
     /**
      * @var string[]|null
      */
@@ -245,6 +248,26 @@ class WPTestCase extends Unit
         self::$coreTestCaseMap[static::class] = $coreTestCase;
 
         return $coreTestCase;
+    }
+
+    public static function isStrictAboutWpdbConnectionId(): bool
+    {
+        return self::$beStrictAboutWpdbConnectionId;
+    }
+
+    public static function beStrictAboutWpdbConnectionId(bool $beStrictAboutWpdbConnectionId): void
+    {
+        self::$beStrictAboutWpdbConnectionId = $beStrictAboutWpdbConnectionId;
+    }
+
+    public static function getWpdbConnectionId(): ?string
+    {
+        return self::$wpdbConnectionId;
+    }
+
+    public static function setWpdbConnectionId(string $wpdbConnectionId): void
+    {
+        self::$wpdbConnectionId = $wpdbConnectionId;
     }
 
     protected function backupAdditionalGlobals(): void